test: 실패하는 테스트 수정
