+++ src/test/java/com/wanted/lunchmapservice/restaurant/config/RestaurantControllerTestConfig.java
@@ -3,10 +3,10 @@ package com.wanted.lunchmapservice.restaurant.config;
 import com.wanted.lunchmapservice.restaurant.mock.RestaurantMock;
 import com.wanted.lunchmapservice.restaurant.service.RestaurantGetService;
 import org.mockito.Mockito;
+import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 
-@Configuration
+@TestConfiguration
 public class RestaurantControllerTestConfig extends AuthTestConfig{
 
     @Bean
